Build the whole project on top of Slate 1.5.0
Instead of bumping our versions from the previous project, we decided
to move the minimum possible from our layout and use everything we can
from the new version of Slate.

We had to customize the following files on the theme:

   * layouts/layout.erb (custom header)
   * layouts/simple.erb (appconsole)
   * stylesheets/_variables.scss (overwriting confs at the bottom)

This commit also adds the ability to run middleman via
Docker (docker-compose), so it's not necessary to configure the local
ruby environment.