Document `--source-only` and add `--push-only`
The `--source-only` already works to only build and not deploy. But
sometimes we want to just push and not build. An example of that is
when you want to use Docker to build the project, avoiding the
installation of the whole Ruby environment in your computer.

This commit adds `--push-only` to add the ability to execute only the
actions related to `git` and avoid the project build.