Backports all `Getting started` work by Lefteris
There was an old branch from Aug 2016 adding a Getting Started section
but it was too old to just merge to the master branch. This commit
backports all the data adding the new changes on top of it.

It splits the `_overview.md` file into three different ones. Most part
of it is now part of the `_sync.md`, but there are some content on
`changelog.md` and `tutorial.md`.