Improve project structure
Since our REST API (still in beta) is a different kind of beast, not
the future of our main API, we must keep a structure to deal with both
docs.

This commit changes the structure to deal with both docs in a way we
can release new versions and still keep the old ones in a structured
way.